+++ srcs/tokenizer/wordexp_tokenizer.c
@@ -6,7 +6,7 @@
 /*   By: niguinti <0x00fi@protonmail.com>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/12/19 06:36:46 by niguinti          #+#    #+#             */
-/*   Updated: 2020/01/30 01:48:17 by niguinti         ###   ########.fr       */
+/*   Updated: 2020/01/30 03:00:21 by niguinti         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -55,7 +55,7 @@ int		lex_match_dquote(char *s, int *anchor, t_lifo *stack)
 	{
 		if (s[*anchor] == '\\' && !esc)
 			esc = 1;
-		else if (wexp_rules[DQU][s[*anchor]])
+		else if (wexp_rules[DQU][(int)(s[*anchor])])
 		{
 			if (!lex_sequence(s, anchor, stack))
 				return (0);
@@ -96,7 +96,7 @@ int		lex_match_command_sub(char *s, int *anchor, t_lifo *stack)
 			if (!lex_match_command_sub(s, anchor, stack))
 				return (0);
 		}
-		if (wexp_rules[BQU][s[*anchor]])
+		if (wexp_rules[BQU][(int)(s[*anchor])])
 		{
 			if (!lex_sequence(s, anchor, stack))
 				return (0);
